# TIFF whole-file I/O speedup 5-10x with multithreading