… src/tiff.imageio/tiffinput.cpp
… src/tiff.imageio/tiffoutput.cpp