@lgritz 2018-03-14
Beware -- merge of this risky change is imminent!